Prefer class-based messages
For any message with a given code, its associated data should have a
fixed, known structure.

Currently, the code and data structure aren't clearly exposed by the
library API. Codes are referenced using string literals, and data
structures are not checked automatically.

As a primary consequence, the result is error-prone, as the burden of
keeping messages consistent falls entirely on the shoulders of the
developer.

Also, codes and data structures can't be easily documented, and library
users are forced to go through the code base and hope not to miss a code
to obtain a list of the messages they'll have to handle.

This commit provides a partial improvement to the situation by
specializing messages according to their code with a dedicated class.
Messages can now be listed and documented in an exhaustive fashion using
standard Ruby documentation tools (such as YARD).

The data structure of messages still can't be easily validated, but now
we have a place to implement such validations (the specialized message class).

This comes at the price of breaking the Messenger API, but we'd better
pay that cost now rather than later.