Remove implicit message codes from exceptions
Sheetah should handle errors in two ways, dependending on when they
occur.

At load time, when the developer is responsible for any error (for
example when a template specification is incorrect), exceptions should
be raised because the error is likely irrecoverable by the system, and
the developer should fix it instead. `Sheetah::Errors::Error` and its
descendants are used to that end.

At runtime, when the user is responsible for any error (for example,
when a cell value doesn't match its expected type), exceptions should
not be raised, because the user has to fix its input, and the developer
can't do anything on its own. The system should recover from such errors
by returning the appropriate message to the user.
`Sheetah::Messaging::Message` are used to that end.

Unfortunately, the implementation confuses these two rules in the
`Backends` part of the library. There, errors are expected at runtime,
due to invalid inputs from the user, and should therefore not be handled
as exceptions. Yet exceptions (`Sheetah::Sheet::Error` and its
descendants) are used to that end , which leads to the problem of
rescuing and turning them into messages.

The code removed in this commit intended to automate this process. It
was mistaken about the two rules described above in a major way: it
implements message codes for all exceptions, including those that are
strictly intended for developers and thus not concerned at all with
messages.

So, this commit:

- limits the automated transformation of exceptions into messages to
  `Sheetah::Sheet::Error`.
- simplifies the implementation of that transformation (because we do
  not need all that complexity after all).

The next step will be to remove `Sheetah::Sheet::Error` altogether and
replace it by messaging.