Remove backends auto-detection
Even though it might be convenient for users not to have to worry about
which backend should handle their input, the way it was implemented was
not deterministic, and actually error-prone.

By the way, auto-detection isn't actually such a big deal for the
library itself, users can add their own helpers on top of the library,
using with their own rules.

So, instead of spending time on a not-so-valuable effort to find a
better implementation, the best compromise is to remove the feature
altogether and keep things simple.