Simplify the CSV backend
On one hand, the backend shouldn't have to deal with filepaths,
encodings, and so on, because they induce way too many edge cases and
pitfalls. Instead, the user should provide an IO-like object they have
correctly configured beforehand for reading.

The backend should on the other hand know how to configure what actually
has to with CSV documents per se: the column or row separators, the
character used for quoting, and so on. For now, a few configurable
options are implemented and we might want to add more if needed.