Disassociate from the upstream project (1/2)
The upstream project (https://github.com/steeple-org/sheetah) and this
one will evolve separately because upstream will likely be too slow to
adapt to the changes planned for the stable 1.0 version.

The first step is to publish an alternate, modified copy of the library.
According to the Apache-2.0 license, the copy includes the original
license and copyright. It is otherwise considered a new, independant
software project, hence the new name.

The next step is to rename the Ruby library itself. This will happen
later, once the currently pending upstream merge requests are integrated
to the project's master branch (to avoid fruitless merge conflicts).