Support Ruby 3+ and drop 2.7
The maintenance status of Ruby 2.7 has been EOL for quite some time now,
so it is legitimate not to support it any longer.

Also, Ruby 3.3 is now in normal maintenance, so we need to support it.

Updating Nokogiri was therefore required, because:

> nokogiri-1.15.4-x86_64-linux requires ruby version < 3.3.dev, >= 2.7

Changing Rubocop's TargetRubyVersion from 2.7 to 3.0 generated a few
offenses. They will be dealt with later, and for now they are simply
ignored.