Validate messages
This commits builds on the previous one
(59d38e41428ab51a722ffcf56b192f7918531161) to implement a simple but
sufficient validation framework for messages.

The validations are meant to help developers write consistent code
regarding messaging, and especially to uphold the invariants of the
specialized message classes.

Validations are enabled by default. They can be disabled either by
setting the SHEETAH_MESSAGING_VALIDATE_MESSAGES environment variable to
the value `"false"`, or by assigning `false` to the corresponding
attribute of `Sheetah::Messaging.config`.

Note that a validation that fails is considered an error from the
developer, not from the user, and is therefore handled as an exception.
Developers are responsible for ensuring the messages the system
generates are consistent, even if those messages are meant to indicate
inconsistencies in the user's input.