Separate concerns between Message and its variants
`Message` is a generic structure able to represent any variant of
messages. As such, it cannot be assigned a specific, fixed class-level
code. Only variants (i.e. subclasses of `Message`) can and should be
assigned such a code. It appears that we have two separate concerns
here:

1. `Message` and its variants should behave alike.
2. variants should provide an API to get/set a class-level code.

The previous implementation provided both with the `Message` class, and
doing so exposed the inconsistency of `Message` having a class-level
method named `::code`. It also required a clumsy hack to provide a code
validation for the variants that wouldn't affect instances of `Message`.

By introducing the `MessageVariant` class, this commit clears the
inconsistency and keeps the `Message` class focused on its goal, and
clarifies the validations.