Clarify attribute dependencies (value,scalar,composite)
A template attribute has exactly one type, and it is either a `Scalar`
or a `Composite`.

A `Scalar` is made of exactly one value, while a `Composite` is made of
one or more values. Both classes represent their value(s) with the
`Value` class.

For each `Value` in a type, there will be exactly one resulting `Column`
in the `Specification` produced by the application of a template config
to the template.

So, to be more precise, `Scalar` or `Composite` are only abstract
wrappers of columns whose types will be compiled later, when a template
config is applied to the template.

This commit clarifies the relationship between those three entities.