Simplify Column
The responsibility of Column is to hold together pieces of information
about a column. As such, it is not supposed to be clever, and there is
no need to have default values for some of its attributes.

The header_pattern of a Column used to be forcefully frozen, in order to
play nice with the internal requirements of Specification, which used to
store its columns in a Hash with header_patterns as keys. It is no
longer the case, thus the force-freeze is removed.