Freeze columns produced by applications of a template
Commit 17e81bb161484ac69529250e58a18c47ac6f756d allowed instances of
`Column` to exist without systematically being frozen. While the change
made sense as far as the `Column` class is concerned, it led to the
fact that columns used in a specification produced by `Template#apply`
stay mutable. This isn't the correct behavior, because such
specifications are meant to be immutable (sheet processors may use them
but may not modifiy them).

This commit fixes the issue by freezing each column before passing it to
a specification in `Template#apply`.