Don't systematically freeze instances of Template & co
It is not always convenient to only be able to produce frozen instances
of a class. For example, it is impossible to stub/mock methods with
RSpec on a frozen object, which can make the task of writing complete
specs more difficult. I therefore generally prefer to initialize mutable
objects, and explicitly freezing them afterwards when necessary.

This commit applies that reasoning to templates and their dependencies.

Templates are meant to live as shared objects in an application, so that
different configs may be applied concurrently to the same template
instance to produce as many specifications. Because they are
concurrently shared, templates should be considered immutable and
freezing them is a good practice.

When it comes to freezing objects, there are different ways to look at
them and their internal state:

- The object does not have any internal state, or that state is already
  frozen (i.e. it composed of booleans, symbols, etc.). Then it is
  enough to only freeze the object itself with its default `#freeze`
  method.
- The object has some mutable internal state. Two possibilities:
  - It is the one that initialized the state, and thus is responsible
    for it (it "owns" that state).
  - It was passed that state from the outside, and thus is not
    responsible for it (it doesn't "own" it).

When a object has and owns some mutable state, then that state should be
frozen when the object itself is frozen.

When a object has but doesn't own some mutable state, then that state
should not be frozen when the object itself is frozen.

This commit enforces those rules as far as `Template` and its related
classes (`Attribute`, etc.) are concerned.