Values of attributes may also be specified as hashes
A value of an attribute used to be specified using a single Symbol, and
yet that symbol could bear two distinct pieces of information: the type
of that value, and whether the value is required.

While it is convenient to have a single literal to write as value when
writing templates, it should really be considered as syntactic sugar,
and should not be the only way to specify values, as it is somewhat
esoteric and not easily extensible.

To improve clarity and extensibility, this commit adds the possibility
to specify such values as hashes instead of symbols.

Note that this commit does not deprecate the historical use of symbols,
which still are a valid way to specify values.