Report coverage in CI runs
A coverage report is now sent to CodeClimate when the test suite runs in
a CI workflow. The report is sent disregarding the success/failure of
running the suite.

The coverage HTML website is now only generated during successful
workflows for the master branch, because only then would we care about
updating the GitHub Pages with the newly generated website.