Report test results in PR annotations
When a test fails in a test suite run, an annotation is added to the PR
(along with its check run).

During a regular workflow, the test suite runs as many time as there are
combinations of (OS, Ruby) in the RSpec job matrix.

As a consequence, for the same failing test, there might be multiple
annotations, one for each failing combination. Having multiple
annotations for the same test might clutter the PR, but it seems
valuable to have precise results about test failures, so we'll see how
it really turns out.

Should the clutter be too much, one would simply have to limit producing
annotations for a limited set of (OS, Ruby) combinations, just like it's
done with coverage reports.