Report code coverage on pull requests from forks
The credentials used in workflows runs for pull requests from forked
projects are read-only and do not have access to the repository secrets.

CC_TEST_REPORTER_ID is therefore not available in such cases.

The solution is to delegate the task of reporting coverage to another
worklow, trigged not by a pull_request event, but another
(workflow_run), which will have access to the repository secrets.

That said, the paambaati/codeclimate-action doesn't seem compatible with
workflow_run-based workflows (it couldn't send coverage data associated
with the correct Git reference), so this commit implements the
report-to-codeclimate job directly using the official CodeClimate reporter.