Harmonize specs and fix Xlsx#each_row
The three backends share the same API (the one required to behave as a
Sheet). To ensure that each variant enforces the shared API invariants,
it is convenient to write their specs with as much shared examples as
possible.

As it happens, sharing specs revealed a bug in the Xlsx backend: it was
the only one not to always return `self` when calling `#each_row` with a
block, and this commit also fixes that inconsistency.