Input values for backends can't be nil
It makes no sense to pass `nil` as the primary argument to a backend at
initialization, whatever the backend may be. We could provide an API to
handle that error explicitly for the benefit of the users of the
backend.

On the other end, passing `nil` as the primary argument is such an
obvious mistake that we prefer to consider it won't happen and that the
burden of ensuring that invariant belongs to the user of the backend,
rather than its implementer.

Should the need for a dedicated API about `nil` inputs arise, we would
then consider it again.