Sheet::open only rescues InputError
Following the work introduced by commit
e092191ae85a2e5be37400c2a02024e11fe4d1b2, this commit clarifies how
errors should be considered when implementing a kind of Sheet.

The idea is that implementers may legitimately raise exceptions that
are not caused by a mistake in the input. Such exceptions should be easy
to implement and rescue, and therefore it is convenient to expose a base
exception class with no extra behavior in the `Sheet` module.
`Sheetah::Sheet::Error` is that class.

On the other hand, because input errors are also handled using
exceptions, this commit refactors the exception class that used to
represent an input error. Being about a mistake related to the details
of some `Sheet`, this exception class naturally inherits from the new
`Sheetah::Sheet::Error`, and is also renamed from `SheetError` to
`InputError` to make the intent clearer.