Simplify error handling in SheetProcessor
`SheetProcessor#call` relies on three distinct and yet somewhat
overlapping error handling mechanisms:

a. returning instances of Result
b. raising `Sheet::InputError` exceptions somewhere in the stack of the `Backends::open` block
c. calling `Failure#unwrap` somewhere in the stack of the `Do()` block

The last one _c_ is especially tricky because the call to
`Failure#unwrap` can happen anywhere in the call stack of the `Do()`
block, and that stack is complex, even more so when we consider the fact
the block eventually yields control the caller of `SheetProcessor#call`.
As a result, relying on _c_ for error handling can get very confusing,
as it obfuscates the source of an error.

Furthermore, and contrary to _b_, _c_ is a generic way to handle
any kind of failures, whereas _b_ is only involved in a precise and
limited scope of errors (those that partake in input errors identified
by Sheet backends). The genericity of _c_ doesn't serve the needs of
`SheetProcessor#call` well, as it expects instances of Result of a
certain shape (empty successes or failures wrapping `#to_message`-able
objects).

This commit therefore replaces the use of _c_ by plain and simple uses
of _a_.