Expose Messenger in backends
The commit e092191ae85a2e5be37400c2a02024e11fe4d1b2 recommended the
abandon of the use of exceptions as a control flow mechanism in backends
facing errors in an input, as well as relying on a messenger to pass
messages instead of using exceptions.

This commit is about the latter.

Indeed, InputError is still used for control flow, because a practical
replacement hasn't been found yet, but it is no longer involved in
passing messages to a messenger.

Backends now receive a messenger at initialization, and thus are free to
pass any message, as many time as they need, to that messenger.

NOTE: backends should probably not touch the scope of their messenger,
as, during the processing of a document, it will most likely be
controlled by other components involved in that processing.