Expose consistent indices for rows/cols
Any (non-header) row is supposed to be identified by an integer-based
index, and any column by a letter-based one. Indices would always start
at 1 or A (respectively), even if the actual tabular document could use
other coordinates for the actual cells.

This commit changes this behavior.

A backend is now required to expose its rows/cols/cells with indices
that are consistent with the actual indices of the underlying document.
In other words, if a cell is located at B4 in XLSX spreadsheet editors,
then it must be exposed with (row: 4, col: B) by the XLSX backend.

NOTE: As far as indices are concerned, header rows and data rows are
handled differently: a header row doesn't have a "row" index, because
it is not considered a row, per se. Only data rows have indices. Now
that rows reflect their actual indices, this behavior may seem
confusing, and might change in a latter commit.