Opt-in sizable rows enumerator
Depending on the internal details of a backend, it might be possible to
know how many rows the document has before enumerating them all.

In such cases, we want an enumerator returned by `#each_row` to return
that number (instead of `nil`) when its `#size` method is called. A use
case for this could be for users of the library to validate that a
document doesn't exceed a certain size before processing it.

That being said, implementers of backends should remember that sizable
enumerators of rows are opt-in. The goal of this library is to be
reasonably efficient with large documents, and it wouldn't make sense to
slurp a whole document into memory only to calculate its size.