Xlsx: ignore empty initial rows and columns
The original intent for the XLSX backend was to consider that, for all
documents, data start at row 1, column A.

Unfortunately, there was a mistake in the implementation and data would
always start at row 1, but not always at column A. In fact, any empty
initial rows would be kept, and any empty initial columns would be
ignored.

This commit chooses to resolve the inconsistency by always ignoring
empty initial rows, like the backend already does with empty initial
columns.

As far as XLSX documents are concerned, it does not seem too far-fetched
to imagine that users may submit valid documents with data not starting
at cell A1. So we choose to support such use cases.