MessageVariant comes with Messaging
It is possible to use the Messaging library without using
MessageVariant. However, the benefits of using MessageVariant are
important, and the class itself is small and doesn't bring extra
dependencies, so insisting on not requiring it by default seems
pointless.

This commit makes it so that requiring messaging.rb is now enough to
require the whole library, including MessageVariant.